Recursively apply base hooks for the Instruction dialog
This fixes a bug where selecting the Instruction op type for a modify hook only applies the base hook instead of looking further to see if that hook also has a base hook, and so on. In some cases, this bug produces an error dialog and prevents selecting instructions introduced by base hooks.